refactor(trade-in): decompose into Valuation/Lifecycle/Query services
Behavior-preserving (write-heavy, jeRefs=0). Facade keeps 19-method surface +
5-arg ctor + delegates; internally constructs 3 sub-services (spec + module
untouched). 4 $transaction moved WHOLE; quickBuy orchestrator intact; PII
encrypt/decrypt + Thai-id validation extracted to a pure helpers util, byte-identical.
tsc 0, tests green, lint 0.
